+++ src/main/java/eventloop/EventLoop.java
@@ -1,4 +1,4 @@
-package event_loop;
+package eventloop;
 
 /**
  * This shall serve as the base for implementing an event loop. An event loop
