+++ src/main/java/eventloop/Executable.java
@@ -1,4 +1,4 @@
-package event_loop;
+package eventloop;
 
 /**
  * The interface for objects that can be run.
