+++ src/main/java/eventloop/Identifiable.java
@@ -1,4 +1,4 @@
-package event_loop;
+package eventloop;
 
 /**
  * An interface for objects that have a id, so that in a sense they could be
