+++ src/main/java/eventloop/IdentifiableExecutable.java
@@ -1,4 +1,4 @@
-package event_loop;
+package eventloop;
 
 /**
  * The interface for objects that have both ids and can be executed.
