+++ src/main/java/eventloop/NestableExecutableObject.java
@@ -1,4 +1,4 @@
-package event_loop;
+package eventloop;
 import java.util.*;
 
 /**
