+++ src/main/java/exceptions/InvalidArgumentException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package exceptions;
 
 /**
  * The exception thrown when the argument does not match the expected.
